+++ mne/__init__.py
@@ -30,7 +30,10 @@ except Exception:
 (__getattr__, __dir__, __all__) = lazy.attach_stub(__name__, __file__)
 
 # initialize logging
-from .utils import set_log_level, set_log_file
+from .utils import _no_filelock, set_log_level, set_log_file
 
-set_log_level(None, False)
-set_log_file()
+# _no_filelock: this single config read happens on every ``import mne``, and locking
+# it would mean importing filelock (~20 ms of asyncio + sqlite3) up front
+with _no_filelock():
+    set_log_level(None, False)
+    set_log_file()
