+++ mne/channels/montage.py
@@ -33,7 +33,6 @@ from ..fixes import _reshape_view
 from ..transforms import (
     Transform,
     _ensure_trans,
-    _fit_matched_points,
     _frame_to_str,
     _quat_to_affine,
     _sph_to_cart,
@@ -1961,6 +1960,8 @@ def compute_dev_head_t(montage):
             f" points in device and head coordinates is required. (Got {len(hpi_dev)}"
             f" points in device and {len(hpi_head)} points in head coordinate systems)"
         )
+
+    from .._transforms_numba import _fit_matched_points
 
     trans = _quat_to_affine(_fit_matched_points(hpi_dev, hpi_head)[0])
     return Transform(fro="meg", to="head", trans=trans)
