+++ mne/conftest.py
@@ -31,10 +31,11 @@ register_assert_rewrite("mne.utils._testing")
 import mne
 from mne import Epochs, create_info, make_fixed_length_epochs, pick_types, read_events
 from mne._fiff.constants import FIFF
+from mne._numba import has_numba
 from mne.channels import read_layout
 from mne.coreg import create_default_subject
 from mne.datasets import testing
-from mne.fixes import _compare_version, has_numba
+from mne.fixes import _compare_version
 from mne.io import RawArray, read_raw_ctf, read_raw_fif, read_raw_nirx, read_raw_snirf
 from mne.utils import (
     Bunch,
