+++ mne/decoding/time_delaying_ridge.py
@@ -11,9 +11,9 @@ from scipy.sparse.csgraph import laplacian
 from sklearn.base import BaseEstimator, RegressorMixin
 from sklearn.utils.validation import check_is_fitted
 
+from .._numba import jit
 from ..cuda import _setup_cuda_fft_multiply_repeated
 from ..filter import next_fast_len
-from ..fixes import jit
 from ..utils import ProgressBar, _check_option, logger, warn
 from ._fixes import _check_n_features_3d, validate_data
