+++ mne/utils/__init__.pyi
@@ -91,6 +91,7 @@ __all__ = [
     "_julian_to_date",
     "_limit_blas_threads",
     "_mask_to_onsets_offsets",
+    "_no_filelock",
     "_on_missing",
     "_open_lock",
     "_parse_verbose",
@@ -289,6 +290,7 @@ from .config import (
     _get_numpy_libs,
     _get_root_dir,
     _get_stim_channel,
+    _no_filelock,
     _open_lock,
     get_config,
     get_config_path,
